feat: universal "Exit Their Economy" loading screen on mobile
Add a shared LoadingScreen component that renders the canonical dark
"Exit Their Economy / Exit The Psyop" loading state, matching the web
app's app/loading.tsx and the mobile design mockup. Wire it as the
app-wide bootstrapping gate in App.tsx (while the stored sign-in session
is being restored) and have the existing per-surface loaders delegate to
it so the loading state is identical everywhere. Loading is intentionally
not theme-toggled per the design spec.

Closes #259
