+++ ctf/packages/mobile/src/features/socketrelay/SocketRelayLoading.tsx
@@ -1,38 +1,8 @@
 import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { LoadingScreen } from '../../components/shared/LoadingScreen';
 
+// Delegates to the universal LoadingScreen so the "Exit Their Economy / Exit The
+// Psyop" loading state is identical everywhere in the app.
 export function SocketRelayLoading() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.line1}>EXIT THEIR ECONOMY</Text>
-      <Text style={styles.line2}>EXIT THE PSYOP</Text>
-    </View>
-  );
+  return <LoadingScreen />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#0F1117',
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 32,
-  },
-  line1: {
-    fontSize: 10,
-    letterSpacing: 2,
-    color: 'rgba(255,255,255,0.22)',
-    textTransform: 'uppercase',
-    fontWeight: '500',
-    marginBottom: 8,
-    textAlign: 'center',
-  },
-  line2: {
-    fontSize: 10,
-    letterSpacing: 2,
-    color: 'rgba(255,255,255,0.22)',
-    textTransform: 'uppercase',
-    fontWeight: '500',
-    textAlign: 'center',
-  },
-});
